finishes user story 1

+++ db/migrate/20190405160509_create_snacks.rb
@@ -2,7 +2,7 @@ class CreateSnacks < ActiveRecord::Migration[5.1]
   def change
     create_table :snacks do |t|
       t.string :name
-      t.integer :price
+      t.money :price
     end
   end
 end
